# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DeguBase – workspace and table management API.",
        "title": "DeguBase API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/workspaces/{wsCode}/tables/{tableCode}/rows/{rowID}/files": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Upload file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "table code",
                        "name": "tableCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "row ID",
                        "name": "rowID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "column code",
                        "name": "colCode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Row"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/tables/{tableCode}/rows/{rowID}/files/{fileID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Download file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "table code",
                        "name": "tableCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "row ID",
                        "name": "rowID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file ID",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/tables/{tableCode}/rows/{rowID}/files/{fileID}/thumbnail": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "image/jpeg"
                ],
                "tags": [
                    "files"
                ],
                "summary": "File thumbnail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "table code",
                        "name": "tableCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "row ID",
                        "name": "rowID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file ID",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "not_found"
                },
                "error": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
...
        "models.Row": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "id": {
                    "type": "integer"
                },
                "table_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}